Hopefully fixed the buildPollChoices curcular dependency
